+++ src/Game/Packet/InPacket.h
@@ -5,7 +5,7 @@
 
 class InPacket : public Packet {
 public:
-	InPacket(unsigned char buffer_in[], unsigned short import_size);
+	InPacket(unsigned char buffer_in[], int import_size);
 
 	template<typename T> T read();
 private:
